fix: Api Pane IA (#32603)
## Description

Fixes the IA in API Response Pane by moving the meta information about
the API call inside the response tab

Fixes #32602

## Automation

/ok-to-test tags="@tag.Datasource"

### :mag: Cypress test results
